Remove dependency on pkg/charts
