+++ viz/tap/pkg/util.go
@@ -8,7 +8,7 @@ import (
 	netPb "github.com/linkerd/linkerd2/controller/gen/common/net"
 	"github.com/linkerd/linkerd2/pkg/k8s"
 	metricsPb "github.com/linkerd/linkerd2/viz/metrics-api/gen/viz"
-	"github.com/linkerd/linkerd2/viz/pkg/api"
+	"github.com/linkerd/linkerd2/viz/pkg"
 	tapPb "github.com/linkerd/linkerd2/viz/tap/gen/tap"
 )
 
@@ -50,7 +50,7 @@ func BuildTapByResourceRequest(params TapRequestParams) (*tapPb.TapByResourceReq
 	if err != nil {
 		return nil, fmt.Errorf("target resource invalid: %s", err)
 	}
-	if !contains(api.ValidTargets, target.Type) {
+	if !contains(pkg.ValidTargets, target.Type) {
 		return nil, fmt.Errorf("unsupported resource type [%s]", target.Type)
 	}
